read access polished

+++ maendeleo-loan-api/src/main/java/co/ke/bank/maendeleo/controllers/LoanApplicationController.java
@@ -33,7 +33,7 @@ public class LoanApplicationController {
 	}
 	
 	@GetMapping(value = "/loan/{memberIdNo}")
-	public ResponseEntity<Object> read(@PathVariable("memberIdNo") final Long memberIdNo){
+	public ResponseEntity<Object> read(@PathVariable("memberIdNo") final int memberIdNo){
 		final List<LoanApplication> response = service.read(memberIdNo);
 		return new ResponseEntity<Object>(response, HttpStatus.OK);
 	}
